fix: addjust route and add authentication

+++ app/Http/Controllers/ProductController.php
@@ -2,19 +2,31 @@
 
 namespace App\Http\Controllers;
 
+use App\Entities\ResponseEntities;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use JWTAuth;
 
 class ProductController extends Controller
 {
     public static function addProduct(Request $request)
     {
-        return Product::addProduct($request->input());
+
+        $user = JWTAuth::parseToken()->authenticate();
+        if ($user->is_admin === 1) {
+            return Product::addProduct($request->input());
+        }
+        return self::cannotAccess();
     }
 
     public static function editProduct(Request $request, $productId)
     {
-        return Product::editProduct($request->input(), $productId);
+
+        $user = JWTAuth::parseToken()->authenticate();
+        if ($user->is_admin === 1) {
+            return Product::editProduct($request->input(), $productId);
+        }
+        return self::cannotAccess();
     }
 
     public static function getProductList(Request $request)
@@ -29,6 +41,18 @@ class ProductController extends Controller
 
     public static function deleteProduct(Request $request, $productId)
     {
-        return Product::deleteProduct($productId);
+        $user = JWTAuth::parseToken()->authenticate();
+        if ($user->is_admin === 1) {
+            return Product::deleteProduct($productId);
+        }
+        return self::cannotAccess();
+    }
+
+    private static function cannotAccess()
+    {
+        $response = new ResponseEntities();
+
+        $response->message = 'Cannot access';
+        return $response;
     }
 }
